Added a PAL stretch option
Stretches the image by X pixel vertically. Meant primarily for fixing bad PAL ports that didn't bother with filling the whole PAL resolution of 288p and instead rendered at 240p but can be used with any game that has black bars (for example Indiana Jones or GoldenEye in both PAL and NTSC).

Also updated the hints for a couple of existing options